# packages/contracts/openapi/school-os-api.yaml
openapi: 3.1.0
info:
  title: School OS API
  version: "1.0"
  description: >
    Versioned public/partner-facing surface of the School OS platform.
    Authoritative implementation lives in apps/platform (Laravel);
    this file is the source of truth for its shape and is consumed by
    packages/shared-types to generate TypeScript bindings. See
    docs/architecture/API.md for the full architecture (auth, pagination,
    idempotency, rate limiting, webhooks, error format).
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /schools/{schoolId}/subjects/{subjectId}:
    get:
      operationId: getSubject
      summary: Fetch one Subject. Requires academics.subjects.view.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { name: subjectId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Subject" } } }
    patch:
      operationId: updateSubject
      summary: Updates a Subject. Requires academics.subjects.manage.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { name: subjectId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Subject" } } }
...
  /schools/{schoolId}/enrollments/{enrollmentId}:
    get:
      operationId: getEnrollment
      summary: Fetch one Enrollment's detail. Requires enrollments.view.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/EnrollmentId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Enrollment" } } }
        "404":
          description: Not found (including a cross-School id -- indistinguishable from nonexistent).
...
  /schools/{schoolId}/enrollment-rollovers/{rollover}:
    get:
      operationId: getEnrollmentRollover
      summary: >
        Rollover Plan detail -- metadata, Mappings, and a durable
        execution summary. Requires enrollments.view AND
        enrollments.rollovers.view. Never embeds the (potentially very
        large) Item collection; see the separate paginated Items
        endpoint.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/RolloverId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/RolloverPlan" } } }
        "404":
          description: Not found (including a cross-School id -- indistinguishable from nonexistent).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: >
        Opaque personal-access token (web/mobile) or partner API key
        (third-party integrations). See docs/architecture/API.md
        ("Authentication").
  parameters:
...
    SchoolId:
      name: schoolId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    EnrollmentId:
      name: enrollmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    RolloverId:
      name: rollover
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Subject:
      type: object
      required: [id, name, code, subjectType, status]
      description: School-wide reference data (section 62).
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        shortName: { type: string, nullable: true }
        subjectType: { type: string, enum: [core, elective, co_scholastic, language, other] }
        academicDepartmentId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [active, inactive] }
    SubjectInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string, maxLength: 255 }
        code: { type: string, maxLength: 32 }
        short_name: { type: string, maxLength: 32, nullable: true }
        subject_type: { type: string, enum: [core, elective, co_scholastic, language, other] }
        academic_department_id: { type: string, format: uuid, nullable: true }
...
    StudentSummary:
      type: object
      required: [id, studentNumber, firstName, status]
      description: >
        List-row shape -- deliberately excludes dateOfBirth (Highly
        Sensitive children's data, docs/security/DATA-CLASSIFICATION.md:
        minimized default visibility in broad list views). See Student
        for the full single-record shape.
      properties:
        id: { type: string, format: uuid }
        studentNumber: { type: string }
        firstName: { type: string }
        middleName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive] }
...
    EnrollmentRef:
      type: object
      required: [id, name, code]
      description: A minimal AcademicYear/Campus/GradeLevel/Section reference embedded in an Enrollment row.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
    EnrollmentSummary:
      type: object
      required: [id, student, academicYear, campus, gradeLevel, section, rollNumber, status, startsOn]
      description: >
        Shared shape for directory rows AND single-record detail (there
        is no broader "full" shape) -- deliberately excludes Student
        dateOfBirth and all Guardian data
        (docs/security/DATA-CLASSIFICATION.md, "minimized default
        visibility"). rollNumber is always a string, never coerced to
        a number (e.g. "007" stays "007").
      properties:
        id: { type: string, format: uuid }
        student: { $ref: "#/components/schemas/StudentSummary" }
        academicYear: { $ref: "#/components/schemas/EnrollmentRef" }
        campus: { $ref: "#/components/schemas/EnrollmentRef" }
        gradeLevel: { $ref: "#/components/schemas/EnrollmentRef" }
        section: { $ref: "#/components/schemas/EnrollmentRef" }
        rollNumber: { type: string }
        status: { type: string, enum: [active, completed, withdrawn, transferred, cancelled] }
        startsOn: { type: string, format: date }
        endsOn: { type: string, format: date, nullable: true }
    Enrollment:
      allOf:
        - { $ref: "#/components/schemas/EnrollmentSummary" }
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    RolloverPlanSummary:
      type: object
      required: [id, sourceAcademicYear, targetAcademicYear, status, configurationVersion, createdAt]
      description: >
        Phase 1B.7E. Directory-row shape -- deliberately omits mappings
        and executionSummary (only present on RolloverPlan detail) to
        keep the directory listing cheap regardless of Plan count.
      properties:
        id: { type: string, format: uuid }
        sourceAcademicYear: { $ref: "#/components/schemas/EnrollmentRef" }
        targetAcademicYear: { $ref: "#/components/schemas/EnrollmentRef" }
        status: { type: string, enum: [draft, validated, executing, completed, completed_with_errors, cancelled] }
        configurationVersion: { type: integer }
        validatedConfigurationVersion: { type: integer, nullable: true }
        createdBy: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        createdAt: { type: string, format: date-time }
        validatedAt: { type: string, format: date-time, nullable: true }
        executionStartedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
    RolloverMapping:
      type: object
      required: [id, sourceGradeLevel, targetGradeLevel, isRepeat]
      properties:
        id: { type: string, format: uuid }
        sourceGradeLevel: { $ref: "#/components/schemas/EnrollmentRef" }
        sourceSection: { allOf: [{ $ref: "#/components/schemas/EnrollmentRef" }], nullable: true }
        targetGradeLevel: { $ref: "#/components/schemas/EnrollmentRef" }
        targetSection: { allOf: [{ $ref: "#/components/schemas/EnrollmentRef" }], nullable: true }
        isRepeat: { type: boolean }
    RolloverSubjectOfferingRef:
      type: object
      required: [id, status]
      description: >
        Phase 1G.4. A minimal, PII-free SubjectOffering reference for
        the subject-mapping workspace -- Subject name/code and academic
        dimensions (GradeLevel/Campus/ElectiveGroup) are non-PII School
        reference/configuration data. `status` is always present so a
        durable mapping's target can be displayed (with an inactive
        warning) even after the Offering itself later became inactive
        (docs/students/PHASE-1G-4-SUBJECT-ROLLOVER-API-UI.md).
      properties:
        id: { type: string, format: uuid }
        subject: { allOf: [{ $ref: "#/components/schemas/EnrollmentRef" }], nullable: true }
        gradeLevel: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        campus: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        status: { type: string, enum: [active, inactive] }
        electiveGroup: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
    RolloverSubjectMapping:
      type: object
      required: [id, sourceSubjectOffering, state]
      description: >
        Phase 1G.4/1G.1. `state` is the three-state model surfaced for
        display: this schema only ever represents a row that EXISTS
        (`mapped` or `omit`) -- the UNCONFIGURED state is the ABSENCE
        of a row, represented separately by
        `RolloverPlan.unmappedSourceSubjectOfferings`, never collapsed
        into this list.
      properties:
        id: { type: string, format: uuid }
        sourceSubjectOffering: { $ref: "#/components/schemas/RolloverSubjectOfferingRef" }
        state: { type: string, enum: [mapped, omit] }
        targetSubjectOffering: { allOf: [{ $ref: "#/components/schemas/RolloverSubjectOfferingRef" }], nullable: true }
...
    RolloverPlan:
      allOf:
        - { $ref: "#/components/schemas/RolloverPlanSummary" }
        - type: object
          required: [isValidatedForCurrentConfiguration, mappings, subjectMappings, unmappedSourceSubjectOfferings, executionSummary]
          properties:
            isValidatedForCurrentConfiguration: { type: boolean }
            mappings: { type: array, items: { $ref: "#/components/schemas/RolloverMapping" } }
            subjectMappings: { type: array, items: { $ref: "#/components/schemas/RolloverSubjectMapping" } }
            unmappedSourceSubjectOfferings:
              type: array
              description: >
                Phase 1G.4. Bounded discovery list of elective source
                SubjectOfferings with current eligible participation in
                this Plan but NO mapping row yet -- what would trigger
                `missing_subject_mapping` in dry-run. Never implies a
                mapping row was created; purely a read-only projection
                (EnrollmentRolloverReadService::unmappedSourceSubjectOfferings()).
              items: { $ref: "#/components/schemas/RolloverSubjectOfferingRef" }
            executionSummary: { $ref: "#/components/schemas/RolloverExecutionSummary" }
...
    RolloverExecutionSummary:
      type: object
      required: [total, succeeded, reconciled, skipped, failed, pending]
      description: >
        Recomputed from persisted Items every time -- never an
        in-memory counter. `planStatus` is present only on start()/
        resume()'s own response (EnrollmentRolloverExecutionService's
        return value), not on the Plan-detail-embedded executionSummary.
      properties:
        total: { type: integer }
        succeeded: { type: integer }
        reconciled: { type: integer }
        skipped: { type: integer }
        failed: { type: integer }
        pending: { type: integer }
        planStatus: { type: string, enum: [draft, validated, executing, completed, completed_with_errors, cancelled] }
...